requirements.txt
django
djangorestframework
django-cors-headers
django-environ
psycopg2-binary
drf-spectacular
drf_spectacular_sidecar
celery[redis]